requirements.txt
absl-py==2.4.0
appnope==0.1.4
asttokens==3.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
contextlib2==21.6.0
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
einops==0.6.1
et_xmlfile==2.0.0
executing==2.2.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
h5py==3.15.1
hf-xet==1.2.0
huggingface-hub==0.36.0
icecream==2.1.3
idna==3.11
imageio==2.28.1
ipykernel==6.25.2
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.8.0
matplotlib-inline==0.2.1
MedPy==0.4.0
ml_collections==0.1.1
monai==1.5.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
nibabel==5.1.0
numpy==1.26.4
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.1.1
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==3.20.3
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.1
PySide6==6.7.3
PySide6_Addons==6.7.3
PySide6_Essentials==6.7.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
setuptools==80.9.0
shiboken6==6.7.3
simpleitk==2.5.3
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tensorboardX==2.6
threadpoolctl==3.6.0
timm==1.0.24
tokenizers==0.19.1
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.65.0
traitlets==5.14.3
transformers==4.41.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.5.0
wheel==0.45.1
